Login via background process
Attempts to log in via a background process instead of in context. This will allow for a longer login process than is currently set for our context, and only attempts to log in one at a time instead of stacking multiple logins if an HTTP call was made multiple concurrent times.