Support multiple paths to a single Tesla Gateway
This adds support to have multiple IP/Ports to calling a single Tesla Gateway, such as if the Gateway is connected to your LAN while the machine running the proxy is also connected to its Wifi network. Will return the first successful response from the gateay, defaulting to the first error otherwise.